AngJS

+++ src/main/java/com/wnycl/dao/TeamDao.java
@@ -5,13 +5,13 @@ import java.util.List;
 import com.wnycl.model.Team;
 
 public interface TeamDao {
-	Team findById(int id);
+	Team findById(long id);
 	
 	Team findByName(String name);
 	
 	void save(Team team);
 	
-	void deleteById(int id);
+	void deleteById(long id);
 	
 	List<Team> findAllTeams();
 }
